+++ file/src/main/java/org/discu2/froum/file/FileApplication.java
@@ -1,0 +1,12 @@
+package org.discu2.froum.file;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FileApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(FileApplication.class, args);
+    }
+}
