+++ file/src/main/java/org/discu2/froum/file/FileController.java
@@ -1,0 +1,9 @@
+package org.discu2.froum.file;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/file")
+public class FileController {
+}
